Fix AES decryption issue
While adding modes of operation the fact that ciphertext of different
ciphers can be stored as arrays of different data types, was omitted.
While trying to adopt it to BlowFish, AES decryption was left broken.
Fix includes additional parameter in ModeOfOperation's encrypt and
decrypt functions' params which tells the size in bytes of message
storage data type